app/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.0
frozenlist==1.4.0
grpcio==1.59.2
grpcio-tools==1.59.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.1
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
multidict==6.0.4
mypy==1.6.1
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
pandas==2.1.2
portalocker==2.8.2
protobuf==4.25.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
qdrant-client==1.6.4
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
Werkzeug==3.0.1
yarl==1.9.2
